+++ CHANGES.md
@@ -2,7 +2,8 @@
 
 **September 6, 2012**
 
-* The html contructor of TextView and HTMLFragment is renamed to fromHTML
+* The content/src constructor of Style is renamed to fromContent/fromSrc, respectively
+* The html constructor of TextView and HTMLFragment is renamed to fromHTML
 * In EasingMotion, duration is renamed to period, repeat parameter is introduced, and mode is removed.
 
 **August 31, 2012**